Car plate page changed to modal, added validation and deleted old unused "icono" css clases
