+++ assets/js/app.js
@@ -6,7 +6,6 @@ import Layout from './containers/Layout/Layout';
 import Login from './containers/Auth/Auth';
 import Home from './containers/Home/Home';
 import Users from './containers/Users/Users';
-import Plate from './containers/PlatePage/Plate';
 import Logout from './containers/Auth/Logout/Logout'
 
 class App extends Component {
@@ -26,7 +25,6 @@ class App extends Component {
                     <Route path="/home" component={Home}/>
                     <Route path='/users' component={Users}/>
                     <Route path='/logout' component={Logout}/>
-                    <Route path='/plate' component={Plate}/>
                     <Redirect from='/' to='/home'/>
                 </Switch>
             </Layout>
