+++ assets/js/index.js
@@ -9,8 +9,7 @@ import App from './app';
 import '../css/index.scss';
 import '../css/colors/index.css';
 import '../css/menu/menu.css';
-import '../css/containers/PlatePage/plate.css';
-import '../css/icons/icons.css';
+import '../css/modal/plateModal.scss';
 
 import rootReducer from './store/reducers/main';
